requirements.txt
Django==1.8.3
django-debug-toolbar==1.3.2
django-filter==0.10.0
djangorestframework==3.2.0
Markdown==2.6.2
Pillow==2.9.0
sqlparse==0.1.16
wheel==0.24.0
dj-database-url==0.3.0
django-postgrespool==0.3.0
gunicorn==19.3.0
psycopg2==2.6
SQLAlchemy==1.0.4
whitenoise==1.0.6
pafy
